Fix nir_op_ffloor and nir_op_ffract
This fixes nir_op_ffloor and nir_op_ffract broken for > INT_MAX values.
These functions (and fceil) are trivial for values > 2^24, as for these values, the floating point precision is less than integral. So fix it by selecting the trivial solution, if the value is too big.

Performance: This introduces a slowdown of approx. 66 % (10 instr. instead of 6).